refactor(lua): rationalize types
Use `size_t` for things that are byte counts, or for passing to APIs
that expect `size_t` params.

For things that are object counts, prefer `unsigned`.

There is one part of this refactoring that might be a bit buggy: in
`commandt_score()` (`score.c`), I added a short circuit where we break
out of the `while` loop if it is impossible for the rest of the needle
to be found in the rest of the haystack:

    while (haystack_idx >= needle_idx) {

The problem with this is that, right after the `while`, we store the
computed mask:

    if (compute_bitmasks) {
        haystack->bitmask = mask;
    }

Evidently, if we short-circuit the loop, then the mask won't actually be
complete. I have a vague recollection that I might have tried a
refactoring like this in the past, and ended up discarding it for
exactly the same reason. Whatever; in the next commit I will plug this
hole.

+++ lua/wincent/commandt/commandt.h
@@ -7,6 +7,7 @@
 #define COMMANDT_H
 
 #include <stdbool.h> /* for bool */
+#include <stddef.h> /* for size_t */
 
 #include "str.h" /* for str_t */
 
@@ -26,13 +27,12 @@ typedef struct {
     /**
      * Number of candidates currently stored in the scanner.
      */
-    // TODO: use size_t for things that are measured in byte sizes.
-    size_t count;
+    unsigned count;
 
     /**
      * Available capacity in the scanner.
      */
-    size_t capacity;
+    unsigned capacity;
 
     /**
      * Counter that increments any time the candidates change.
@@ -60,7 +60,7 @@ typedef struct {
      * Limit the number of returned results. Must be non-zero.
      */
     unsigned limit;
-    int threads;
+    unsigned threads;
 
     /**
      * Note that the matcher doesn't take ownership of the `needle` (ie. it
@@ -70,11 +70,11 @@ typedef struct {
      * and friends.
      */
     const char *needle;
-    unsigned long needle_length;
+    size_t needle_length;
     long needle_bitmask;
 
     const char *last_needle;
-    unsigned long last_needle_length;
+    size_t last_needle_length;
 } matcher_t;
 
 #endif
